+++ packages/frontend/src/components/common/Table/ScrollableTable/TableHeader.tsx
@@ -1,8 +1,9 @@
 import { Draggable } from '@hello-pangea/dnd';
-import { isField } from '@lightdash/common';
+import { FieldType, isField } from '@lightdash/common';
 import { Tooltip } from '@mantine/core';
 import { flexRender } from '@tanstack/react-table';
-import React, { type FC } from 'react';
+import isEqual from 'lodash/isEqual';
+import React, { useEffect, type FC } from 'react';
 import {
     TABLE_HEADER_BG,
     Th,
@@ -15,12 +16,49 @@ import { HeaderDndContext, HeaderDroppable } from './HeaderDnD';
 
 interface TableHeaderProps {
     minimal?: boolean;
+    showSubtotals?: boolean;
 }
 
-const TableHeader: FC<TableHeaderProps> = ({ minimal = false }) => {
+const TableHeader: FC<TableHeaderProps> = ({
+    minimal = false,
+    showSubtotals = true,
+}) => {
     const { table, headerContextMenu, columns } = useTableContext();
     const HeaderContextMenu = headerContextMenu;
     const currentColOrder = React.useRef<Array<string>>([]);
+
+    useEffect(() => {
+        if (showSubtotals) {
+            const groupedColumns = columns
+                .filter((col) => {
+                    const item = col.meta?.item;
+                    return item && isField(item)
+                        ? item.fieldType === FieldType.DIMENSION
+                        : false;
+                })
+                .map((col) => col.id);
+
+            const sortedColumns = table
+                .getState()
+                .columnOrder.reduce<string[]>((acc, sortedId) => {
+                    return groupedColumns.includes(sortedId)
+                        ? [...acc, sortedId]
+                        : acc;
+                }, [])
+                // The last dimension column essentially groups rows for each unique value in that column.
+                // Grouping on it would result in many useless expandable groups containing just one item.
+                .slice(0, -1);
+
+            if (!isEqual(sortedColumns, table.getState().grouping)) {
+                table.setGrouping(sortedColumns);
+            }
+        } else {
+            if (table.getState().grouping.length > 0) {
+                table.resetGrouping();
+            }
+        }
+    }, [showSubtotals, columns, headerContextMenu, table]);
+
     if (columns.length <= 0) {
         return null;
     }